fix: make font install smarter

+++ install.sh
@@ -107,16 +107,18 @@ _bsd() {
 }
 
 _fonts() {
-    $sudo apt install fontconfig
-    cd ~
-    wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/Meslo.zip
-    mkdir -p .local/share/fonts
-    unzip Meslo.zip -d .local/share/fonts
-    cd .local/share/fonts
-    rm *Windows*
-    cd ~
-    rm Meslo.zip
-    fc-cache -fv
+    if [[ ! -d $HOME/.local/share/fonts ]]; then
+        $sudo apt install fontconfig
+        cd ~
+        wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/Meslo.zip
+        mkdir -p .local/share/fonts
+        unzip Meslo.zip -d .local/share/fonts
+        cd .local/share/fonts
+        rm *Windows*
+        cd ~
+        rm Meslo.zip
+        fc-cache -fv
+    fi
 }
 
 _debian() {
